backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

psycopg2-binary==2.9.9
requests==2.31.0
paramiko==3.4.0
redis==5.0.1
email-validator==2.1.0.post1


routeros-api==0.17.0

